[UPDATE] fix bugs, remove 'TODO' comment and update minor modules

+++ include/ncurses_display.h
@@ -4,14 +4,18 @@
 
 #include <curses.h>
 
+#include <string>
+#include <vector>
+
 #include "./process.h"
 #include "./system.h"
 
 namespace NCursesDisplay {
-void Display(System& system, int n = 10);
+void Display(System& system, int16_t n = 10);
 void DisplaySystem(System& system, WINDOW* window);
-void DisplayProcesses(std::vector<Process>& processes, WINDOW* window, int n);
+void DisplayProcesses(std::vector<Process>& processes, WINDOW* window,
+                      int16_t n);
 std::string ProgressBar(float percent);
 };  // namespace NCursesDisplay
 
-#endif
+#endif  // NCURSES_DISPLAY_H
